Chore(Refactored index.js to include tests
