+++ jest.config.js
@@ -1,0 +1,11 @@
+export default {
+    preset: 'ts-jest',
+    testEnvironment: 'jest-environment-node',
+    testMatch: ['**/**/*.test.js'],
+    verbose: true,
+    forceExit: true,
+    transform: {},
+    // clearMocks: true
+
+
+}
